B #6114: Fix NETTX and NETRX host monitoring (#2494)

+++ src/host/HostMonitoringTemplate.cc
@@ -30,7 +30,7 @@ using namespace std;
 
 int CapacityMonitoring::from_template(const Template &tmpl)
 {
-    int value;
+    unsigned long value;
     if (tmpl.get("FREECPU", value))
     {
         add("FREE_CPU", value);
@@ -55,7 +55,7 @@ int CapacityMonitoring::from_template(const Template &tmpl)
 
 int SystemMonitoring::from_template(const Template &tmpl)
 {
-    int value;
+    unsigned long value;
     if (tmpl.get("CPUSPEED", value))
     {
         add("CPU_SPEED", value);
